// src/types.ts
export interface Cliente { // Define la interfaz Cliente que refleja la entidad del backend
    id: number;   // Identificador único del cliente        
    nombre: string;  // Nombre del cliente
    apellido: string; // Apellido del cliente
    telefono: string; // Telefono del cliente
    email: string;  // Correo electrónico del cliente   
    direccion: string; // Direccion del cliente
  }
  